build: Fail loudly if bash variables are unset, or a pipeline errors
These bash options increase the robustness of bash scripts, increasing
the chances that errors are spotted, instead of passing silently.

errexit was already enabled (by bash -e) if the scripts were invoked as

path/to/script.sh

However bash -e would not take effect if the script was invoked as

bash path/to/script.sh

The long form of the options is used to serve as an memory aide for
readers unfamiliar with bash options.